Moved gapply() to its own file
